refactor: hooks namespace
