+++ src/JsonApiDotNetCore/Hooks/EntityDiff.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using JsonApiDotNetCore.Models;
 
-namespace JsonApiDotNetCore.Services
+namespace JsonApiDotNetCore.Hooks
 {
     public interface IEntityDiff<TEntity> where TEntity : class, IIdentifiable
     {
