+++ src/JsonApiDotNetCore/Hooks/Traversal/RelationshipsFromPreviousLayer.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using JsonApiDotNetCore.Internal;
 using JsonApiDotNetCore.Models;
 
-namespace JsonApiDotNetCore.Services
+namespace JsonApiDotNetCore.Hooks
 {
    internal interface IRelationshipsFromPreviousLayer
     {
