+++ src/JsonApiDotNetCore/Hooks/Traversal/RootNode.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using JsonApiDotNetCore.Internal;
 using JsonApiDotNetCore.Models;
 
-namespace JsonApiDotNetCore.Services
+namespace JsonApiDotNetCore.Hooks
 {
     internal class RootNode<TEntity> : IEntityNode where TEntity : class, IIdentifiable
     {
