+++ src/JsonApiDotNetCore/Hooks/UpdatedRelationshipHelper.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using JsonApiDotNetCore.Internal;
 using JsonApiDotNetCore.Models;
 
-namespace JsonApiDotNetCore.Services
+namespace JsonApiDotNetCore.Hooks
 {
     public interface IUpdatedRelationshipHelper { }
